Add Read At Indicator (#485)
* Add Read At Indicator

This PR should add a new checkmark read at indicator if the message has
been read and is the last message in the conversation. The popup does
not look the best but can adjust with any feedback. There is a case in
the fixtures where a message can have two notifications, is that a use
case to worry about?

* Add a new feature `message_read_indicator`
* Add a new component `Messages::ReadAt`

* Change indicator to show for whole conversation

* Add feature check for read indicator

* Add Read At Indicator

This PR should add a new checkmark read at indicator if the message has
been read and is the last message in the conversation. The popup does
not look the best but can adjust with any feedback. There is a case in
the fixtures where a message can have two notifications, is that a use
case to worry about?

* Add a new feature `message_read_indicator`
* Add a new component `Messages::ReadAt`

* Change indicator to show for whole conversation

* Fix bad rebase

* Change `message_read_indicator` to based on admin or dev

* Localize messages for conversation/read_indicator

* Change assert_not to refute

* Fix linting issue

* Fix indentation

* Change method to `latest_message_read_by_other_recipient?`

* Update read indicator

* Add turbo frame tag for message read indicator

* Remove sutbbing from test

* [skip ci] Update CHANGELOG

Co-authored-by: Joe Masilotti <joe@masilotti.com>